:label: [rename] Video Player Multi와 Single로 분리 (HH-323)
